Added items in a meta sense, and items in a real sense, also now you can use items! Items!

+++ js/bubbles_game/Entity.pde
@@ -1,9 +1,9 @@
 class Entity {
   PVector position;
   PVector velocity;
-  int radius;
+  float radius;
   
-  Entity(PVector p, PVector v, int r) {
+  Entity(PVector p, PVector v, float r) {
     this.position = p;
     this.velocity = v;
     this.radius = r;
